requirements.txt
streamlit
requests
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
